First Day
Back in highschool I learnt a few things about code in computer class but I don't even remember of we got to use anything that wasn't scratch or cmd.exe.

This is the start of my Journey in coding, I have read and listened to severa opinions. I'm sure many people document their coding journey, this is mainly to save my process some where on the go and to take a look back if I really end up liking this. It seems I will.